==> Precard_eGFR_DTII_Longitudinal_Models/02a_Precard_LCM_Dataset_Preparation_Report.Rmd <==
---
title: "Precard dataset presentation"
author: "Tasos Mangelis"
date: "`r Sys.Date()`"
output: 
  rmarkdown::html_vignette:
    toc: true

vignette: >
  %\VignetteIndexEntry{Vignette Title}
  %\VignetteEngine{knitr::rmarkdown}
  %\VignetteEncoding{UTF-8}
---

```{r setup, include = FALSE}
knitr::opts_chunk$set(
  collapse = TRUE,
  comment = "#>"
)
library(pander)
library(knitr)
library(summarytools)
library(gtsummary)
library(ggplot2)

library(lcsm)
library(stringr)
```

# Outline of the dataset

The data consist of 15 individual extractions of hospital databases in .csv files.

These have been put together in the following datasets:

| 21805 patients            | T1 diabetes   | T2 diabetes    |
|---------------------------|---------------|----------------|
| Baseline                  | 5826 patients | 15979 patients |
| eGFR longitudinal         | 108246 obs    | 489344 obs     |
| eGFR longitudinal grouped | 29586 obs     | 92945 obs      |
| eGFR Survival             | 5826 patients | 15979 patients |

# Baseline dataset

The `Kidney baseline dataset` is put together by the raw data that we got from hospital databases.

The data were standardised according to the `first available measurement of eGFR` and all the baseline values of the other variables were the first available within a 2 year span from the eGFR baseline.

    eGFR dataset 
      first available measurement identification: eGFR baseline
        subsequent merging with other variables and identification of 2-year
        nearest value
          merging of all data in one dataset

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8}

DiagrammeR::grViz("digraph {
  graph [layout = dot, rankdir = TB]
  
  node [shape = rectangle]        
  rec1 [label = 'Step 1. Data from hospital database']
  rec2 [label = 'Step 2. Standarisation to the first available eGFR']
  rec3 [label = 'Step 3. Merging with all other datasets']
  rec4 [label = 'Step 4. Data transformation']
  rec5 [label = 'Long data with baseline']
  rec6 [label = 'Long data - yearly']
  rec7 [label = 'Survival data']
  # edge definitions with the node IDs
  rec1 -> rec2 -> rec3 -> rec4 -> rec5 
  rec4 -> rec6
  rec4 -> rec7
  }")

```

## Features

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis'}
names(precard_T2%>%
  select(-c(R50_ID, test_date.bsl, test_year.bsl, test_date, test_year, Creatinine, YOBirth, YODeath, DM_TYPE_ID)))%>%
  kable()
```

## Dataset head

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis'}
head(precard_T2%>%
  select(-c(test_date.bsl, test_year.bsl, test_date, test_year, Creatinine, YOBirth, YODeath, DM_TYPE_ID)), n = 10)%>%
  kable()

```

## Missing values

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8}
DataExplorer::plot_missing(precard_long)

```

## Descriptives

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis'}
print(dfSummary(precard_T2%>%select(-c(R50_ID, test_date.bsl, test_date, DM_TYPE_ID, DIAGNOSIS_YEAR, YOBirth, YODeath))), method = "render")
```

## Figures

### Figure 1: eGFR over Gender and Ethnicity

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8}
ggplot(precard_T2) +
  aes(x = "", y = eGFR, colour = Ethnicity_groups) +
  geom_boxplot(fill = "#112446") +
  scale_color_hue(direction = 1) +
  theme_minimal() +
  facet_wrap(vars(Gender))
```

### Figure 2: Age over Gender and Ethnicity

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8}
ggplot(precard_T2) +
  aes(x = "", y = Age, colour = Ethnicity_groups) +
  geom_boxplot(fill = "#112446") +
  scale_color_hue(direction = 1) +
  theme_minimal() +
  facet_wrap(vars(Gender))
```

### Figure 3: eGFR over Age and grouped by ethnicity

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8, fig.show='hold'}
  
ggplot(precard_T2) +
  aes(x = Age, y = eGFR) +
  geom_point(shape = "circle", size = 1.5, colour = "#112446") +
  theme_minimal() +
  facet_wrap(vars(Ethnicity_groups))
  
  

```

### Figure 4: eGFR plotted against import clinical predictors

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8, fig.show='hold'}
ggplot(precard_T2) +
  aes(x = Hba1c, y = eGFR) +
  geom_point(shape = "circle", size = 1.5, colour = "#112446") +
  theme_minimal()
  
ggplot(precard_T2) +
  aes(x = ACR, y = eGFR) +
  geom_point(shape = "circle", size = 1.5, colour = "#112446") +
  theme_minimal()
  
ggplot(precard_T2) +
  aes(x = BMI, y = eGFR) +
  geom_point(shape = "circle", size = 1.5, colour = "#112446") +
  theme_minimal()


```

### Figure 5: eGFR stratified by retinopathy and maculopathy status

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8}
ggplot(precard_T2) +
  aes(x = RM, y = eGFR) +
  geom_boxplot(fill = "#112446") +
  theme_minimal()

```

# Longitudinal dataset

Following the `Baseline dataset`, I calculated the difference of all eGFR available measurements, and time difference from baseline.

## Dataset head

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis'}
head(precard_long%>%
  select(-c(R50_ID, test_date.bsl, test_date, Creatinine, YOBirth, YODeath, DM_TYPE_ID)), n = 10)%>%
  kable()

```

## Missing data

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=8}

DataExplorer::plot_missing(precard_longT2_p)

```

## Figures

### Violin plots

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=10}

ggplot(precard_longT2_group, aes(as.factor(timeDiff), egfr_year)) +
  geom_violin() +
  geom_boxplot(width = 0.1, outlier.colour = "blue") +
  theme_classic()

```

### Overlaid individual trajectories

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=10}
plot.vars<- c("0", "1", "2", "3", "4", "5", "6", "7", "8", "9", 
"10", "11", "12", "13", "14")
plot_trajectories(data = precard_longT2_pc,
                  id_var = "R50_ID", 
                  var_list = plot.vars,
                  xlab = "Follow-up years", ylab = "eGFR (ml/min)",
                  connect_missing = FALSE, 
                  random_sample_frac = 0.2)

```

### Separate individual trajectories

```{r echo=FALSE, message=FALSE, warning=FALSE, results='asis', fig.height=6, fig.width=10}
# Create plot for construct x
# Add facet_wrap() function from ggplot2
plot_trajectories(data = precard_longT2_pc,
                  id_var = "R50_ID", 
                  var_list = plot.vars,
                  xlab = "Follow-up years", ylab = "eGFR (ml/min)",
                  connect_missing = F, 
                  random_sample_frac = 0.057, 
                  title_n = T) +
  facet_wrap(~R50_ID)
```
